Move func def parseFile to where it is called
